requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0
bitsandbytes>=0.39.0
llama-cpp-python>=0.2.90

# RAG and retrieval
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
rank-bm25>=0.2.2

# Data processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# UI and utilities
rich>=13.7.1
typer>=0.12.0
pydantic>=2.0.0
pyyaml>=6.0

# Optional polishing model
# flan-t5-small will be loaded optionally

# Existing dependencies from simple system
orjson>=3.10.0
requests>=2.32.0




